# backend/swagger/flowfi.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowFi API",
    "version": "1.0.0",
    "description": "API documentation for FlowFi - Real-time payment streaming on Stellar\n\n## Performance & Caching\nThe API implements caching for frequently accessed endpoints, such as claimable amount calculations. \n- **Claimable Cache TTL**: 5 seconds\n- **Invalidation**: Automatically cleared when a withdrawal event occurs.\n\n## Sandbox Mode\n\nFlowFi API supports sandbox mode for testing without affecting production data.\n\n**Enable Sandbox Mode:**\n- Header: `X-Sandbox-Mode: true`\n- Query Parameter: `?sandbox=true`\n\n**Sandbox Features:**\n- Isolated database (separate from production)\n- All responses include `_sandbox` metadata\n- Response headers include `X-Sandbox-Mode: true`\n- Safe for testing and development\n\nSee [Sandbox Mode Documentation](../docs/SANDBOX_MODE.md) for details.",
    "contact": {
      "name": "FlowFi Team",
      "url": "https://github.com/LabsCrypt/flowfi"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/v1",
      "description": "Development server (v1)"
    },
    {
      "url": "https://api.flowfi.io/v1",
      "description": "Production server (v1)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JSON Web Token issued by /v1/auth/verify after completing the SEP-10 challenge flow."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Alias for BearerAuth — used by route-level security annotations."
      },
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT — the token subject must match ADMIN_PUBLIC_KEY."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              },
              "details": {
                "type": "array",
                "description": "Structured validation issues when applicable",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/v1/admin/indexer/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get indexer status",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Indexer status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/indexer/reset": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reset indexer lastProcessedLedger (supports dry-run preview)",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dryRun",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "If true, return the projected reset scope without mutating state."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ledger"
                ],
                "properties": {
                  "ledger": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "lastLedger": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ledger value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/indexer/replay": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Replay events from a given ledger (supports dry-run preview)",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from_ledger",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "dryRun",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "If true, return the projected replay scope without mutating state."
          }
        ],
        "responses": {
          "200": {
            "description": "Dry-run preview of the replay scope"
          },
          "202": {
            "description": "Replay started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "replayingFrom": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid from_ledger value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/indexer/dead-letter": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List quarantined indexer events",
        "description": "Paginated, filterable view of Soroban events the indexer failed to\nprocess. Each entry carries the raw payload and the most recent error,\nwith `attempts` recording how many times processing has been tried.\n",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "in": "query",
            "name": "ledgerSequence",
            "schema": {
              "type": "integer"
            },
            "description": "Return only events from this ledger"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "eventType",
            "schema": {
              "type": "string",
              "example": "stream_created"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated dead-letter records"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin access required"
          }
        }
      }
    },
    "/v1/admin/indexer/dead-letter/replay-all": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Replay every pending dead-letter event",
        "description": "Replays all pending records sequentially, oldest ledger first, so events\nfor the same stream apply in their original order.\n",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Per-record replay outcomes plus an aggregate summary"
          }
        }
      }
    },
    "/v1/admin/indexer/dead-letter/{id}/replay": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Replay a single quarantined event",
        "description": "Re-injects the payload through the indexer pipeline. On success the\ndead-letter row is deleted and the corresponding Stream / StreamEvent\nrecords are created; on failure `attempts` is incremented and\n`lastAttemptAt` / `errorMessage` are refreshed.\n",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event replayed and removed from the dead-letter table"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Dead-letter event not found"
          },
          "422": {
            "description": "Replay ran but failed, or the payload is undecodable"
          }
        }
      }
    },
    "/v1/admin/indexer/dead-letter/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Permanently discard a quarantined event",
        "description": "Deletes an unrecoverable record. The discard is written to the\napplication log with the acting admin key so the loss of an on-chain\nevent remains auditable.\n",
        "security": [
          {
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record discarded"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Dead-letter event not found"
          }
        }
      }
    }
  }
}